Done
Pretty much all thats left is to spell check and verify the text.
