+++ FitnessApp/src/Site.css
@@ -119,7 +119,16 @@ h5 {
   --power-1-out: cubic-bezier(0.25, 0.46, 0.45, 0.94);
   --power-2-out: cubic-bezier(0.17, 0.67, 0.83, 0.67);
   /*  */
-  /* Temp */
+  /* Img opacity layers */
+  /*  */
+  --img-opacity-low: rgba(2, 24, 23, 0.2);
+  --img-opacity-medium: rgba(2, 24, 23, 0.4);
+  --img-opacity-high: rgba(2, 24, 23, 0.6);
+
+  /*  */
+  /*  */
+  /* Temp to support old backend or app design thats not the frontpage*/
+  /*  */
   /*  */
   /* Opacity versions */
   --opacity-c1-normal: hsl(155, 100%, 44%);
