+++ FitnessApp/src/view/landing/about/about.css
@@ -165,6 +165,18 @@
   border-radius: 40px;
 }
 
+.about-section-cards-right-image:before {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: var(--img-opacity-low);
+  border-radius: 40px;
+  z-index: 1;
+}
+
 @media (max-width: 1500px) {
   .about-section-cards {
     flex-direction: column;
@@ -177,7 +189,7 @@
   }
 }
 
-@media (max-width: 1500px) {
+@media (max-width: 1200px) {
   .about-section-header-text-para {
     flex-direction: column;
   }
