+++ FitnessApp/src/view/landing/hero/hero.css
@@ -34,6 +34,7 @@
   align-items: flex-start;
   justify-content: center;
   z-index: 5;
+  opacity: 0;
 }
 /*  */
 /* Land section text */
