+++ FitnessApp/src/view/landing/login/text.js
@@ -1,0 +1,1 @@
+export const para = `Hi if your intrested you can login or create an account it's free. Otherwise have a great day :)`;
