--- FitnessApp/src/view/landing/newLogin/text.js
@@ -1,1 +1,0 @@
-export const para = `Hi if your intrested you can login or create an account it's free. Otherwise have a great day :)`;
