+++ FitnessApp/src/view/landing/preview/preview.css
@@ -79,6 +79,18 @@
   border-radius: 40px;
 }
 
+.preview-section-images-container-item:before {
+  content: "";
+  position: absolute;
+  top: 0;
+  right: 0;
+  width: 100%;
+  height: 100%;
+  background-color: var(--img-opacity-low);
+  border-radius: 40px;
+  z-index: 15;
+}
+
 .preview-section-images-container-item.active {
   width: 50%;
 }
